[TESTS] Include apt cookbook into platform run list
While is already `apt-get update` was part of kitchen-dokken's
`intermediate_instructions`, this is not sufficient to keep lists up to date.

These intermediate instructions are executed during container *build* time,
thus outdated, if it is used a few weeks later.

Therefore, add the `apt` cookbook to the platform's run list to refresh packages
at the beginning of the chef run.