[BUGFIX] Don't write jenkins-plugin-last-state file to file_cache_path
As described in someara/kitchen-dokken#118, the contents of
Chef::Config[:file_cache_path] are persistent on the host's file system.

This leads to the unexpected effect that Jenkins is not reliably restarted
after plugin installation, if the previous run of kitchen-dokken had the
same plugins installated.

Therefore, put the file into a directory inside the container, which is
reliably destroyed together with the container.